Refactors lib helpers modularization
