+++ lib/helpers/index.js
@@ -1,0 +1,10 @@
+import './misc'
+import {
+  randomBetween,
+  randomBetweenWithString
+} from './numbers'
+
+export const numbers = {
+  randomBetween,
+  randomBetweenWithString
+}
