+++ lib/index.js
@@ -1,6 +1,5 @@
 import { createFile } from './create-file'
 import createModel from './create-model'
-import { getLocation } from './helpers'
 
 const outputTypes = {
   // Generates the Json output
